購物車功能完成 crud

+++ src/redux/cartSlice.jsx
@@ -1,8 +1,20 @@
 import { createSlice } from '@reduxjs/toolkit'
 import axios from 'axios'
 
+//ReadMe
+// dispatch(fetchUserCart(memberId))
+// dispatch(addItemToCart(memberId, 102))
+// dispatch(addQuantity(memberId, 295, 1))
+// dispatch(minusQuantity(memberId, 295, 2))
+// dispatch(deleteItem(memberId, 293))
+// dispatch(emptyCart(memberId))
+
 const initialState = {
   cartItem: [],
+  cartTotalPrice: 0,
+  cartInfoModal: false,
+  deleteInfoModal: false,
+  clickItem: { itemName: 0, itemSid: 0 },
 }
 
 const cartSlice = createSlice({
@@ -10,30 +22,51 @@ const cartSlice = createSlice({
   initialState,
   reducers: {
     setCartItem(state, action) {
-      console.log(action.payload, 'payload')
       state.cartItem = action.payload
+    },
+    setCartTotalPrice(state, action) {
+      state.cartTotalPrice = action.payload
+    },
+    setCartInfoModalOpen(state, action) {
+      state.cartInfoModal = action.payload
+    },
+    setCartInfoModalClose(state, action) {
+      state.cartInfoModal = action.payload
+    },
+    setCartDeleteInfoModalOpen(state, action) {
+      state.deleteInfoModal = action.payload
+    },
+    setCartDeleteInfoModalClose(state, action) {
+      state.deleteInfoModal = action.payload
+    },
+    setClickItem(state, action) {
+      state.clickItem = { ...action.payload }
     },
   },
 })
 
 const fetchCartData = async (member_id) => {
   const res = await axios.get(`http://localhost:3003/cart/api/${member_id}`)
-
-  console.log(res.data.rows, 'res')
-
+  console.log(res, 'res')
   if (!res.data) {
     throw new Error('fetch fail!')
   }
+
   return res.data.rows
 }
 
 //得到購物車資料
 export const fetchUserCart = (member_id) => {
-  console.log(member_id)
   return async (dispatch) => {
     try {
       const cartData = await fetchCartData(member_id)
       dispatch(cartAction.setCartItem(cartData))
+
+      const totalPrice = cartData.reduce(
+        (init, item) => init + +item.product_price * +item.quantity,
+        0
+      )
+      dispatch(cartAction.setCartTotalPrice(totalPrice))
     } catch (error) {
       console.log(error)
     }
@@ -42,18 +75,29 @@ export const fetchUserCart = (member_id) => {
 //新增商品進購物車
 export const addItemToCart = (member_id, product_id) => {
   return async (dispatch) => {
-    const res = await axios.post('http://localhost:3003/cart/addItem', {
-      member_id,
-      product_id,
-    })
-
-    if (res.data[0].affectedRows !== 1) {
-      throw new Error('新增失敗 請重新再試!!')
-    }
-
     try {
-      const cartData = await fetchUserCart(member_id)
+      if (!member_id) {
+        dispatch(cartAction.setCartInfoModalOpen(true))
+        throw new Error('請先登入在進行操作')
+      }
+
+      const res = await axios.post('http://localhost:3003/cart/addItem', {
+        member_id,
+        product_id,
+      })
+
+      if (res.data.affectedRows !== 1) {
+        throw new Error(res.data)
+      }
+
+      const cartData = await fetchCartData(member_id)
       dispatch(cartAction.setCartItem(cartData))
+
+      const totalPrice = cartData.reduce(
+        (init, item) => init + +item.product_price * +item.quantity,
+        0
+      )
+      dispatch(cartAction.setCartTotalPrice(totalPrice))
     } catch (error) {
       throw new Error(error)
     }
@@ -63,6 +107,11 @@ export const addItemToCart = (member_id, product_id) => {
 //刪除購物車物品
 export const deleteItem = (member_id, sid) => {
   return async (dispatch) => {
+    if (!member_id) {
+      dispatch(cartAction.setCartInfoModalOpen(true))
+      throw new Error('請先登入在進行操作')
+    }
+
     const res = await axios.delete(
       `http://localhost:3003/cart/delete_item/${member_id}/${sid}`
     )
@@ -70,10 +119,15 @@ export const deleteItem = (member_id, sid) => {
     if (res.data[0].affectedRows !== 1) {
       throw new Error('刪除失敗 請重新再試!!')
     }
-
     try {
       const cartData = await fetchCartData(member_id)
       dispatch(cartAction.setCartItem(cartData))
+
+      const totalPrice = cartData.reduce(
+        (init, item) => init + +item.product_price * +item.quantity,
+        0
+      )
+      dispatch(cartAction.setCartTotalPrice(totalPrice))
     } catch (error) {
       throw new Error(error)
     }
@@ -83,19 +137,30 @@ export const deleteItem = (member_id, sid) => {
 //商品數量 +1
 export const addQuantity = (member_id, sid, quantity) => {
   return async (dispatch) => {
-    const res = await axios.put(`http://localhost:3003/cart/plus_quantity`, {
-      quantity,
-      sid,
-      member_id,
-    })
-
-    if (res.data[0].affectedRows !== 1) {
-      throw new Error('商品新增失敗 請重新再試!!')
-    }
-
     try {
+      if (!member_id) {
+        dispatch(cartAction.setCartInfoModalOpen(true))
+        throw new Error('請先登入在進行操作')
+      }
+
+      const res = await axios.put(`http://localhost:3003/cart/plus_quantity`, {
+        quantity,
+        sid,
+        member_id,
+      })
+
+      if (res.data[0].affectedRows !== 1) {
+        throw new Error('商品新增失敗 請重新再試!!')
+      }
+
       const cartData = await fetchCartData(member_id)
       dispatch(cartAction.setCartItem(cartData))
+
+      const totalPrice = cartData.reduce(
+        (init, item) => init + +item.product_price * +item.quantity,
+        0
+      )
+      dispatch(cartAction.setCartTotalPrice(totalPrice))
     } catch (error) {
       throw new Error(error)
     }
@@ -105,19 +170,30 @@ export const addQuantity = (member_id, sid, quantity) => {
 //商品數量-1
 export const minusQuantity = (member_id, sid, quantity) => {
   return async (dispatch) => {
-    const res = await axios.put(`http://localhost:3003/cart/minus_quantity`, {
-      member_id,
-      quantity,
-      sid,
-    })
-
-    if (!res.data[0].affectedRows !== 1) {
-      throw new Error('商品減少失敗 請重新再試!!')
-    }
-
     try {
+      if (!member_id) {
+        dispatch(cartAction.setCartInfoModalOpen(true))
+        throw new Error('請先登入在進行操作')
+      }
+
+      const res = await axios.put(`http://localhost:3003/cart/minus_quantity`, {
+        member_id,
+        quantity,
+        sid,
+      })
+
+      if (res.data[0].affectedRows !== 1) {
+        throw new Error('商品減少失敗 請重新再試!!')
+      }
+
       const cartData = await fetchCartData(member_id)
       dispatch(cartAction.setCartItem(cartData))
+
+      const totalPrice = cartData.reduce(
+        (init, item) => init + +item.product_price * +item.quantity,
+        0
+      )
+      dispatch(cartAction.setCartTotalPrice(totalPrice))
     } catch (error) {
       throw new Error(error)
     }
@@ -126,19 +202,29 @@ export const minusQuantity = (member_id, sid, quantity) => {
 
 //清空購物車
 export const emptyCart = (member_id) => {
-  console.log(`http://localhost:3003/cart/empty_cart/${member_id}`)
   return async (dispatch) => {
-    const res = await axios.delete(
-      `http://localhost:3003/cart/empty_cart/${member_id}`
-    )
-
-    if (res.data[0].affectedRows !== 1) {
-      throw new Error('清空購物車失敗 請重新再試!!')
-    }
-
     try {
+      if (!member_id) {
+        dispatch(cartAction.setCartInfoModalOpen(true))
+        throw new Error('請先登入在進行操作')
+      }
+
+      const res = await axios.delete(
+        `http://localhost:3003/cart/empty_cart/${member_id}`
+      )
+
+      if (res.data[0].affectedRows !== 1) {
+        throw new Error('清空購物車失敗 請重新再試!!')
+      }
+
       const cartData = await fetchCartData(member_id)
       dispatch(cartAction.setCartItem(cartData))
+
+      const totalPrice = cartData.reduce(
+        (init, item) => init + +item.product_price * +item.quantity,
+        0
+      )
+      dispatch(cartAction.setCartTotalPrice(totalPrice))
     } catch (error) {
       throw new Error(error)
     }
@@ -146,4 +232,11 @@ export const emptyCart = (member_id) => {
 }
 
 export const cartAction = cartSlice.actions
+export const {
+  setCartInfoModalOpen,
+  setCartInfoModalClose,
+  setCartDeleteInfoModalOpen,
+  setCartDeleteInfoModalClose,
+  setClickItem,
+} = cartSlice.actions
 export default cartSlice.reducer
